+++ config/ember-try.js
@@ -6,18 +6,10 @@ module.exports = async function() {
   return {
     scenarios: [
       {
-        name: 'ember-lts-3.4',
+        name: 'ember-3.11',
         npm: {
           devDependencies: {
-            'ember-source': '~3.4.0'
-          }
-        }
-      },
-      {
-        name: 'ember-lts-3.8',
-        npm: {
-          devDependencies: {
-            'ember-source': '~3.8.0'
+            'ember-source': '~3.11.0'
           }
         }
       },