+++ tests/integration/test-helpers/calendar-select-test.js
@@ -17,8 +17,8 @@ module('Test Support | Helper | calendarSelect', function(hooks) {
           @selected={{selected4}}
           @onSelect={{action (mut selected4) value="date"}}
           @onCenterChange={{action (mut center4) value="date"}} as |calendar|>
-          <calendar.nav/>
-          <calendar.days/>
+          <calendar.Nav/>
+          <calendar.Days/>
         </PowerCalendar>
       </div>
     `);
@@ -40,8 +40,8 @@ module('Test Support | Helper | calendarSelect', function(hooks) {
           @selected={{selected4}}
           @onSelect={{action (mut selected4) value="date"}}
           @onCenterChange={{action (mut center4) value="date"}} as |calendar|>
-          <calendar.nav/>
-          <calendar.days/>
+          <calendar.Nav/>
+          <calendar.Days/>
         </PowerCalendar>
       </div>
     `);